+++ client/order/src/main/java/com/live_commerce/order/application/dto/response/OrderGetOneResponse.java
@@ -10,7 +10,7 @@ public record OrderGetOneResponse(
         UUID productId,
         UUID userId,
         Integer productQuantity,
-        Long productTotalPrice,
+        Double productTotalPrice,
         String requirement,
         OrderStatus status,
         UUID broadcastId,
