+++ client/order/src/main/java/com/live_commerce/order/infrastructure/client/PaymentReadyRequestDto.java
@@ -5,6 +5,6 @@ import java.util.UUID;
 
 public record PaymentReadyRequestDto(
         UUID orderId,
-        Long amount,
+        double amount,
         String itemName
 ) {}
